requirements.txt
kafka-python
six
mysql.connector
apache-airflow